Cleanup data transfer between devices.
