--- app/src/full/java/io/homeassistant/companion/android/settings/views/SettingsWearLandingView.kt
@@ -1,96 +1,0 @@
-package io.homeassistant.companion.android.settings.views
-
-import android.content.Intent
-import android.net.Uri
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Button
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
-import androidx.compose.material.Scaffold
-import androidx.compose.material.Text
-import androidx.compose.material.TopAppBar
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.HelpOutline
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import io.homeassistant.companion.android.util.wearDeviceName
-import io.homeassistant.companion.android.common.R as commonR
-
-@Composable
-fun SettingWearLandingView(
-    deviceName: String,
-    isAuthed: Boolean,
-    navigateFavorites: () -> Unit,
-    loginWearOs: () -> Unit
-) {
-    val context = LocalContext.current
-    Scaffold(
-        topBar = {
-            TopAppBar(
-                title = { Text(stringResource(commonR.string.wear_settings)) },
-                actions = {
-                    IconButton(onClick = {
-                        val intent = Intent(
-                            Intent.ACTION_VIEW,
-                            Uri.parse(WEAR_DOCS_LINK)
-                        )
-                        context.startActivity(intent)
-                    }) {
-                        Icon(
-                            Icons.Filled.HelpOutline,
-                            contentDescription = stringResource(id = commonR.string.help)
-                        )
-                    }
-                }
-            )
-        }
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(start = 20.dp, top = 10.dp, end = 20.dp)
-        ) {
-            Text(
-                text = stringResource(id = commonR.string.manage_favorites_device, deviceName),
-                textAlign = TextAlign.Center
-            )
-            if (isAuthed) {
-                Button(
-                    onClick = navigateFavorites,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(top = 10.dp, end = 10.dp)
-                ) {
-                    Text(text = stringResource(commonR.string.set_favorites_on_device))
-                }
-            } else {
-                Button(
-                    onClick = loginWearOs,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(top = 10.dp, end = 10.dp)
-                ) {
-                    Text(text = "Login Wear OS Device")
-                }
-            }
-        }
-    }
-}
-
-@Preview
-@Composable
-private fun PreviewSettingWearLandingView() {
-    SettingWearLandingView(
-        deviceName = wearDeviceName,
-        isAuthed = true,
-        navigateFavorites = {},
-        loginWearOs = {}
-    )
-}
