+++ app/src/main/AndroidManifest.xml
@@ -257,12 +257,12 @@
             android:configChanges="orientation|screenSize|keyboardHidden" />
 
         <activity
-            android:name=".settings.SettingsWearActivity"
+            android:name=".settings.wear.SettingsWearActivity"
             android:parentActivityName=".settings.SettingsActivity"
             android:configChanges="orientation|screenSize" />
 
         <activity
-            android:name=".settings.views.SettingsWearMainView"
+            android:name=".settings.wear.views.SettingsWearMainView"
             android:parentActivityName=".settings.SettingsActivity"
             android:configChanges="orientation|screenSize" />
