+++ wear/src/main/java/io/homeassistant/companion/android/phone/PhoneSettingsListener.kt
@@ -1,8 +1,9 @@
 package io.homeassistant.companion.android.phone
 
 import android.content.Intent
-import android.net.Uri
 import android.util.Log
+import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
+import com.fasterxml.jackson.module.kotlin.readValue
 import com.google.android.gms.wearable.DataClient
 import com.google.android.gms.wearable.DataEvent
 import com.google.android.gms.wearable.DataEventBuffer
@@ -25,7 +26,6 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
-import org.json.JSONArray
 import javax.inject.Inject
 
 @AndroidEntryPoint
@@ -42,20 +42,37 @@ class PhoneSettingsListener : WearableListenerService(), DataClient.OnDataChange
 
     private val mainScope: CoroutineScope = CoroutineScope(Dispatchers.Main + Job())
 
+    private val objectMapper = jacksonObjectMapper()
+
     companion object {
         private const val TAG = "PhoneSettingsListener"
-    }
 
-    override fun onCreate() {
-        super.onCreate()
-        Log.d(TAG, "We have favorite message listener")
+        private const val KEY_UPDATE_TIME = "UpdateTime"
+        private const val KEY_IS_AUTHENTICATED = "isAuthenticated"
+        private const val KEY_FAVORITES = "favorites"
     }
 
     override fun onMessageReceived(event: MessageEvent) {
         Log.d(TAG, "Message received: $event")
-        if (event.path == "/send_home_favorites") {
-            val nodeId = event.sourceNodeId
-            sendHomeFavorites(nodeId)
+        if (event.path == "/requestConfig") {
+            sendPhoneData()
+        }
+    }
+
+    private fun sendPhoneData() = mainScope.launch {
+        val currentFavorites =
+            AppDatabase.getInstance(applicationContext).favoritesDao().getAll() ?: listOf()
+        val putDataRequest = PutDataMapRequest.create("/config").run {
+            dataMap.putLong(KEY_UPDATE_TIME, System.nanoTime())
+            dataMap.putBoolean(KEY_IS_AUTHENTICATED, integrationUseCase.isRegistered())
+            dataMap.putString(KEY_FAVORITES, objectMapper.writeValueAsString(currentFavorites))
+            setUrgent()
+            asPutDataRequest()
+        }
+
+        Wearable.getDataClient(this@PhoneSettingsListener).putDataItem(putDataRequest).apply {
+            addOnSuccessListener { Log.d(TAG, "Successfully sent /config to device") }
+            addOnFailureListener { e -> Log.e(TAG, "Failed to send /config to device", e) }
         }
     }
 
@@ -68,15 +85,14 @@ class PhoneSettingsListener : WearableListenerService(), DataClient.OnDataChange
                         "/authenticate" -> {
                             login(DataMapItem.fromDataItem(item).dataMap)
                         }
-                        "/save_home_favorites" -> {
-                            val data = getHomeFavorites(DataMapItem.fromDataItem(item).dataMap)
-                            Log.d(TAG, "onDataChanged: Received home favorites: $data")
-                            saveFavorites()
+                        "/updateFavorites" -> {
+                            saveFavorites(DataMapItem.fromDataItem(item).dataMap)
                         }
                     }
                 }
             }
         }
+        dataEvents.release()
     }
 
     private fun login(dataMap: DataMap) = mainScope.launch {
@@ -94,67 +110,20 @@ class PhoneSettingsListener : WearableListenerService(), DataClient.OnDataChange
             )
         )
 
+        sendPhoneData()
+
         val intent = HomeActivity.newInstance(applicationContext)
         intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK
         startActivity(intent)
     }
 
-    private fun getHomeFavorites(map: DataMap): String {
-        map.apply {
-            return getString("favorites", "")
-        }
-    }
-
-    private fun sendHomeFavorites(nodeId: String) = mainScope.launch {
-        Log.d(TAG, "sendHomeFavorites to: $nodeId")
-        val currentFavorites = AppDatabase.getInstance(applicationContext).favoritesDao().getAll()
-        val list = emptyList<String>().toMutableList()
-        for (favorite in currentFavorites!!) {
-            list += listOf(favorite.id)
-        }
-        val jsonArray = JSONArray(list.toString())
-        val jsonString = List(jsonArray.length()) {
-            jsonArray.getString(it)
-        }.map { it }
-
-        Log.d(TAG, "new list: $jsonString")
-        val putDataRequest = PutDataMapRequest.create("/home_favorites").run {
-            dataMap.putBoolean("isAuthenticated", integrationUseCase.isRegistered())
-            dataMap.putString("favorites", jsonString.toString())
-            setUrgent()
-            asPutDataRequest()
-        }
-
-        Wearable.getDataClient(this@PhoneSettingsListener).putDataItem(putDataRequest).apply {
-            addOnSuccessListener { Log.d(TAG, "Successfully sent favorites to device") }
-            addOnFailureListener { e ->
-                Log.e(TAG, "Failed to send favorites to device", e)
-            }
-        }
-    }
-
-    private fun saveFavorites() {
-        Log.d(TAG, "Finding existing favorites")
+    private fun saveFavorites(dataMap: DataMap) {
+        val favoritesIds: List<String> =
+            objectMapper.readValue(dataMap.getString(KEY_FAVORITES, "[]"))
         val favoritesDao = AppDatabase.getInstance(applicationContext).favoritesDao()
-        mainScope.launch {
-            Wearable.getDataClient(applicationContext).getDataItems(Uri.parse("wear://*/save_home_favorites"))
-                .addOnSuccessListener {
-                    Log.d(TAG, "Found existing favorites: ${it.count}")
-                    it.forEach { dataItem ->
-                        val data = getHomeFavorites(DataMapItem.fromDataItem(dataItem).dataMap).removeSurrounding("[", "]").split(", ").toList()
-                        Log.d(
-                            TAG,
-                            "Favorites: $data"
-                        )
-                        favoritesDao.deleteAll()
-                        if (data.isNotEmpty()) {
-                            data.forEachIndexed { index, s ->
-                                favoritesDao.add(Favorites(s, index))
-                            }
-                        }
-                    }
-                    it.release()
-                }
+        favoritesDao.deleteAll()
+        favoritesIds.forEachIndexed { index, entityId ->
+            favoritesDao.add(Favorites(entityId, index))
         }
     }
 }
