+++ wear/src/main/res/values/wear.xml
@@ -4,7 +4,6 @@
     <string-array name="android_wear_capabilities">
         <item>authentication_token</item>
         <item>verify_wear_app</item>
-        <item>send_home_favorites</item>
-        <item>save_home_favorites</item>
+        <item>sends_config</item>
     </string-array>
 </resources>
